feat: redirect to report detail page after report creation
