+++ src/app/reports/[reportId]/components/ReportDetail/index.module.scss
@@ -1,0 +1,6 @@
+.root {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
